Manually fix category of BuildPermission type.
The type category of BuildPermission was set as Enum for some reason,
even though there was never a corresponding enum element. The category
has been updated to "DataType". The only referring member was removed
fairly early, so relatively few builds are affected.

This only affects legacy json dumps, which are allowed to be modified to
improve accuracy.